+++ CHANGELOG.md
@@ -21,6 +21,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.2.0] - 2026-02-11
+
+### Added
+
+- ical-data-specification.md
+- Experimental tests
+- Tool to create the ical-generator-schema.json
+
+### Changed
+
+- Updated README.md
+
+
 ## [0.1.0] - 2026-02-09
 
 ### Added